expose encoder error
